refactor google service
